midicontroller: Make appearance of channel filters more similar
There are two different kinds of channel filter in the midicontroller
module:

 - "channel", for mappings (in the "list" and "layout" tabs)
 - "channelfilter", for note output (in the savedata/"triangle" menu)

So far, the channelfilter menu has a slightly different appearance than
the channel menu, because it uses numbers padded to two digits (i.e. 01).

This commit replaces the padding "0" with a space to give them the a
more similar appearance.